Move remaining files at root-arch directories into package-arch directories
