+++ feature/ontology/src/commonMain/kotlin/org/pointyware/commonsense/feature/ontology/di/OntologyModule.kt
@@ -5,17 +5,17 @@ import org.koin.core.module.Module
 import org.koin.core.module.dsl.bind
 import org.koin.core.module.dsl.singleOf
 import org.koin.dsl.module
-import org.pointyware.commonsense.feature.ontology.category.data.CategoryDataSource
-import org.pointyware.commonsense.feature.ontology.category.data.CategoryRepository
-import org.pointyware.commonsense.feature.ontology.category.data.CategoryRepositoryImpl
-import org.pointyware.commonsense.feature.ontology.category.data.CategorySqlDataSource
-import org.pointyware.commonsense.feature.ontology.category.interactors.CreateNewCategoryUseCase
-import org.pointyware.commonsense.feature.ontology.category.interactors.CreateNewConceptUseCase
-import org.pointyware.commonsense.feature.ontology.category.interactors.GetSelectedCategoryUseCase
-import org.pointyware.commonsense.feature.ontology.category.interactors.GetSelectedConceptUseCase
-import org.pointyware.commonsense.feature.ontology.category.viewmodels.CategoryEditorViewModel
-import org.pointyware.commonsense.feature.ontology.category.viewmodels.CategoryEditorViewModelImpl
-import org.pointyware.commonsense.feature.ontology.category.viewmodels.CategoryExplorerViewModel
+import org.pointyware.commonsense.feature.ontology.data.CategoryDataSource
+import org.pointyware.commonsense.feature.ontology.data.CategoryRepository
+import org.pointyware.commonsense.feature.ontology.data.CategoryRepositoryImpl
+import org.pointyware.commonsense.feature.ontology.data.CategorySqlDataSource
+import org.pointyware.commonsense.feature.ontology.interactors.CreateNewCategoryUseCase
+import org.pointyware.commonsense.feature.ontology.interactors.CreateNewConceptUseCase
+import org.pointyware.commonsense.feature.ontology.interactors.GetSelectedCategoryUseCase
+import org.pointyware.commonsense.feature.ontology.interactors.GetSelectedConceptUseCase
+import org.pointyware.commonsense.feature.ontology.viewmodels.CategoryEditorViewModel
+import org.pointyware.commonsense.feature.ontology.viewmodels.CategoryEditorViewModelImpl
+import org.pointyware.commonsense.feature.ontology.viewmodels.CategoryExplorerViewModel
 import org.pointyware.commonsense.feature.ontology.data.ArrangementController
 import org.pointyware.commonsense.feature.ontology.data.ConceptEditorController
 import org.pointyware.commonsense.feature.ontology.data.ConceptEditorControllerImpl
