+++ feature/ontology/src/commonMain/kotlin/org/pointyware/commonsense/feature/ontology/interactors/CreateNewConceptUseCase.kt
@@ -1,9 +1,9 @@
-package org.pointyware.commonsense.feature.ontology.category.interactors
+package org.pointyware.commonsense.feature.ontology.interactors
 
 import org.pointyware.commonsense.core.common.Uuid
 import org.pointyware.commonsense.feature.ontology.Concept
 import org.pointyware.commonsense.feature.ontology.IndependentConcept
-import org.pointyware.commonsense.feature.ontology.category.data.CategoryRepository
+import org.pointyware.commonsense.feature.ontology.data.CategoryRepository
 import org.pointyware.commonsense.feature.ontology.data.ConceptEditorController
 
 /**
